feat: 添加设置样式
